requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pinecone-client==3.1.0
python-dotenv==1.0.1
transformers==4.38.2
torch==2.2.1
pillow==10.2.0
numpy==1.26.4
streamlit==1.31.1
requests==2.31.0